# sutra/contracts/sutra.openapi.yaml
openapi: 3.0.3
info:
  title: sutra
  description: The single API surface — CLI, web UI, and external agents all speak this contract.
  version: 0.2.0
paths:
...
  /issues/{issueId}/documents:
    get:
      operationId: listIssueDocuments
      summary: List documents currently tied to an issue
      description: An issue lists its currently tied documents (AC-doc-list-by-issue).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
      responses:
        '200':
          description: Tied documents.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
...
  /projects/{projectId}/documents:
    post:
      operationId: createDocument
      summary: Create a document, optionally from a template
      description: >-
        Files a document under the project, optionally tied to an issue, readable
        back with title and content (AC-doc-create). With template_id set, the
        document's first version starts with the template's content
        (AC-template-instantiate); otherwise content is required. Records a
        `doc.version-saved` event for the initial version, plus a `doc.linked`
        event when created with `issue` set — both stamped with `author` as
        the actor.
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewDocument'
      responses:
        '201':
          description: Document created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listProjectDocuments
      summary: List a project's documents
      description: Documents file under their project (AC-doc-create) — the project-side listing the doc database scenarios read back.
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '200':
          description: The project's documents.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{documentId}:
    get:
      operationId: getDocument
      summary: Get a document, latest version by default
      description: Requesting without a version yields the latest (AC-doc-current).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - name: version
          in: query
          description: A specific version number; omit for the current version.
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: The document and the resolved version's content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentView'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{documentId}/versions:
    post:
      operationId: saveDocVersion
      summary: Save a new document version
      description: >-
        Appends a new immutable DocVersion with author, number, and timestamp;
        prior versions stay readable, no in-place edits (AC-doc-versioning).
        Invoked by the document view's save action (ACT-save-doc).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewDocVersion'
      responses:
        '201':
          description: New version saved and made current.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocVersion'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listDocVersions
      summary: List a document's version history
      description: All versions in order with author and time (AC-doc-history).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
      responses:
        '200':
          description: Version history, oldest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DocVersion' }
        '404':
          $ref: '#/components/responses/NotFound'
  /doc-versions/{docVersionId}:
    get:
      operationId: getDocVersion
      summary: Get a specific document version
      description: >-
        A single DocVersion, including its content, addressable independent
        of the parent document — e.g. to render a doc deliverable pinned by
        a Review (AC-review-web).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocVersionIdPath'
      responses:
        '200':
          description: The document version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocVersion'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{documentId}/diff:
    get:
      operationId: diffDocVersions
      summary: Diff two versions of a document
      description: Any two versions can be diffed (AC-doc-history).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - name: from
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
        - name: to
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: The diff between the two versions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocDiff'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{documentId}/issue:
    post:
      operationId: linkDocumentToIssue
      summary: Tie an existing document to an issue
      description: >-
        An existing document can be tied to an issue after creation
        (AC-doc-link); records a `doc.linked` event (subject: the newly
        linked issue).
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issue, actor]
              properties:
                issue: { type: string, format: uuid }
                actor: { type: string, format: uuid }
      responses:
        '200':
          description: Document with the issue tied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: unlinkDocumentFromIssue
      summary: Untie a document from its issue
      description: >-
        Untied after creation (AC-doc-link); records a `doc.unlinked` event.
      tags: [documents]
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: actor
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Document with the issue untied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # -------------------------------------------------------------- templates
  /templates:
    post:
      operationId: createTemplate
      summary: Create a doc template
      description: Templates can be created, listed, updated, and removed by unique name (AC-template-crud).
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewDocTemplate'
      responses:
        '201':
          description: Template created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocTemplate'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listTemplates
      summary: List doc templates
      tags: [templates]
      responses:
        '200':
          description: Templates.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DocTemplate' }
  /templates/{templateId}:
    get:
      operationId: getTemplate
      summary: Get a doc template
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
      responses:
        '200':
          description: The template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateTemplate
      summary: Update a doc template
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocTemplate'
      responses:
        '200':
          description: Updated template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteTemplate
      summary: Remove a doc template
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '204':
          description: Template removed.
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------------- threads
...
components:
  parameters:
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-supplied key that makes this mutating operation safe to retry
        (CON-idempotent-mutations): replaying the same key returns the
        original response verbatim and has no second effect — no duplicate
        resource, no second state transition, no second side effect.
      schema: { type: string }
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IssueIdPath:
      name: issueId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DocumentIdPath:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    TemplateIdPath:
      name: templateId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DocVersionIdPath:
      name: docVersionId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: >-
        The request is invalid — e.g. an actor or assignee identity id that
        names no known identity (AC-identity-referenced), or a status outside
        the five defined values (AC-status-invalid). The API deals only in
        identity ids; resolving a human-typed handle to an id is the CLI's job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: >-
        The request conflicts with current state — e.g. closing without an
        approved review, a blocking or ancestry cycle, a colliding import, a
        write against an archived project, or a review-anchored comment or
        verdict naming a stale revision (superseded by a resubmission since
        the reviewer last saw it) — rejected atomically, never applied to
        content the reviewer did not see.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [message, code]
      allOf:
        - anyOf:
            - not:
                type: object
                properties:
                  code: { type: string, enum: [unique-violation] }
                required: [code]
            - type: object
              required: [conflicts]
              properties:
                conflicts: { type: array, minItems: 1, items: {} }
          description: A unique-violation always names at least one colliding resource in conflicts — a schema-valid response can never omit what clients need to act on the collision.
      properties:
        message:
          type: string
          description: Human-readable explanation of the rejection.
        code:
          type: string
          enum: [relation-exists, ancestry-cycle, blocking-cycle,
                 project-archived, expected-status-mismatch,
                 expected-revision-mismatch, expected-verdict-event-mismatch,
                 expected-base-mismatch, expected-default-head-mismatch,
                 expected-subtree-revision-mismatch, review-close-used,
                 review-consumed, review-not-approved, open-children,
                 missing-approval, uuid-collision, malformed-import,
                 unique-violation, bad-request, not-found]
          description: >-
            Stable machine-readable cause, for deterministic client
            classification — a closed enum, so schema-valid responses cannot
            invent values. unique-violation covers every uniqueness
            constraint (duplicate identity handles, template names, label
            names), with the colliding resource named in conflicts. Clients distinguishing transient races from
            permanent conditions (e.g. kriya's parenting reconciliation) rely
            on this field, never on message text.
        conflicts:
          type: array
          description: Ids of the resources this request conflicts with, if applicable.
          items: { type: string }
...
    Document:
      type: object
      required: [id, project, title]
      properties:
        id: { type: string, format: uuid }
        project: { type: string, format: uuid }
        issue: { type: string, format: uuid, nullable: true }
        title: { type: string }
        current_version: { type: string, format: uuid, nullable: true }
    DocumentView:
      description: A document together with its resolved version's content (latest by default).
      allOf:
        - $ref: '#/components/schemas/Document'
        - type: object
          required: [version]
          properties:
            version: { $ref: '#/components/schemas/DocVersion' }
    NewDocument:
      type: object
      description: >-
        Exactly one of template_id or content must be given: template_id
        seeds the first version from the template's content
        (AC-template-instantiate); otherwise content is the first version's
        content directly. Both set is rejected.
      required: [title, author]
      properties:
        title: { type: string }
        issue: { type: string, format: uuid }
        template_id: { type: string, format: uuid }
        content: { type: string }
        author:
          type: string
          format: uuid
          description: >-
            Identity recorded as the author of the first DocVersion, and as
            the actor of the creation-time events (`doc.version-saved`, and
            `doc.linked` when created with `issue` set).
      oneOf:
        - required: [template_id]
          not:
            required: [content]
        - required: [content]
          not:
            required: [template_id]
    DocVersion:
      type: object
      required: [id, document, number, content, author, created]
      properties:
        id: { type: string, format: uuid }
        document: { type: string, format: uuid }
        number: { type: integer }
        content: { type: string }
        author: { type: string, format: uuid }
        created: { type: string, format: date-time }
    NewDocVersion:
      type: object
      required: [content, author]
      properties:
        content: { type: string }
        author: { type: string, format: uuid }
    DocDiff:
      type: object
      required: [from, to, diff]
      properties:
        from: { type: integer }
        to: { type: integer }
        diff:
          type: string
          description: Unified diff between the `from` and `to` version contents.
    DocTemplate:
      type: object
      required: [id, name, content]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        content: { type: string }
    NewDocTemplate:
      type: object
      required: [name, content]
      properties:
        name: { type: string }
        content: { type: string }
    UpdateDocTemplate:
      type: object
      properties:
        name: { type: string }
        content: { type: string }
...